updated start_dp so that we can run it on pbs clusters

+++ stop_dp.sh
@@ -1,4 +1,14 @@
 #!/bin/bash
 
-scancel `cat jobid.fit`
-scancel `cat jobid.best`
+if [ `which scancel` ]; then
+    scancel `cat jobid.fit`
+    scancel `cat jobid.best`
+fi
+if [ `which qdel` ]; then
+    echo "qdel" 
+    cat jobid.fit
+    cat jobid.best
+    qdel `cat jobid.fit`
+    sleep 5
+    qdel `cat jobid.best`
+fi
